Make read-your-writes internals private
