+++ fdbclient/SnapshotCache.h
